task_2_mongodb/requirements.txt
pymongo==4.9.1
python-dotenv==1.0.1

